add comment condition to deleteUser func

+++ src/app/pages/post/post.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { PostService } from 'src/app/services/post.service';
 import { Post } from './post';
+import { CommentService } from 'src/app/services/comment.service';
 
 @Component({
   selector: 'app-post',
@@ -8,11 +9,15 @@ import { Post } from './post';
   styleUrls: ['./post.component.css'],
 })
 export class PostComponent {
-  constructor(private postService: PostService) {}
+  constructor(private postService: PostService, private commentService:CommentService) {}
   postList: Post[] = this.postService.getPosts();
 
   deleteUser(id:number){
-    this.postService.deleteUser(id);
+    if(this.commentService.calculateCommentNumber(id)===0){
+      this.postService.deleteUser(id);
+    } else {
+      alert("Gönderiye ait yorum bulunmaktadır, silme işlemi gerçekleştirilemez!")
+    }
   }
   showUserDetails(){
     console.log("Detay sayfası")
